Updated the model package
