+++ library-service/src/model/dto/bookDto.go
@@ -1,0 +1,14 @@
+package dto
+
+import ()
+
+type BookDtos []*BookDto
+
+type BookDto struct {
+	ID            uint   `json:"id"`
+	Title         string `json:"title"`
+	Author        string `json:"author"`
+	PublishedDate string `json:"published_date"`
+	ImageUrl      string `json:"image_url"`
+	Description   string `json:"description"`
+}
